requirements.txt
# Google AI API
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
# xgboost>=1.7.0  # Optional - install separately if needed
# lightgbm>=4.0.0  # Optional - install separately if needed

# Deep Learning (optional - install separately if needed)
# torch>=2.0.0
# transformers>=4.30.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0

# API and HTTP
requests>=2.31.0
aiohttp>=3.8.0

# Configuration
pyyaml>=6.0 